restructure views
